feat: adds ability to use guards on login route and interceptors on cb
Due to a new options strucutre, it is possible to now use interceptors on the OAuth callback URL,
which is useful when it comes to things like needing to add a cookie to the response, without having
to use the `res` object in the controller or the service. Guards are also now available for the
login URL to allow devs to keep users from hitting the URL if they are already logged in. The
options objects for the Google and GitHub OAuth flows have also been updated to match the options
allowed by the OAuth authoirites.

BREAKING CHANGE: The `OauthModuleOptions` have been updated in the `controller` and `service`
objects. In the `controller` object, each route is now a `path` and `guards` or `interceptors`
property, and the `service` now is specific to each provider.