fix: allows for the use of a function that returns an observable
Previously, a function that returned an observable passed to the `provide` property would result in
an error from Nest about the inability to convert a circular JSON structure. Now, the service
factory will check the return of the method. If a `pipe` property exists, it is assumed to be an
observable, and returned directly, otherwise the `from` operator is used to create the observable
for the rest of the function to finish properly.