feat: allows for decorator usage and access to req in `provide`
With the new API there is now access to the `req` object in the `provide` method, meaning that devs
can now assign values to `req._headers` or `req._cookies` and use packages similar to
`@nestjsplus/cookies` for setting cookie values or dynamic headers. This should be the last major
breaking change. Only updates from here should be the addition of providers and the better typings

BREAKING CHANGE: the `provide` method now has a different parameter type, from a simple object to a
complex one.